requirements.txt
Django==1.2.3
South==0.7.3
distribute==0.6.14
django-taggit==0.9.1
docutils==0.7
django-extensions==0.6
